commit 3 20/2/2024
0122

+++ src/main/java/com/venturenix/cmc/payload/response/EventGroupResponse.java
@@ -9,8 +9,8 @@ public class EventGroupResponse {
   private String token;
   private String type = "Bearer";
   private Long id; 
-  private Integer groupid;
-  private Integer eventid;
+  private Long groupid;
+  private Long eventid;
   private String status;
   
   private LocalDateTime createddate;
@@ -25,7 +25,7 @@ public class EventGroupResponse {
   private Integer updatedby;
   
   
-  public EventGroupResponse(String accessToken, Long id, Integer groupid, Integer eventid, String name, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
+  public EventGroupResponse(String accessToken, Long id, Long groupid, Long eventid, String name, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
     this.token = accessToken;
     this.id = id;
     this.groupid = groupid;
@@ -62,19 +62,19 @@ public class EventGroupResponse {
     this.id = id;
   }
 
-  public Integer getGroupid() {
+  public Long getGroupid() {
     return groupid;
   }
 
-  public void setGroupid(Integer groupid) {
+  public void setGroupid(Long groupid) {
     this.groupid = groupid;
   }  
 
-  public Integer getEventid() {
+  public Long getEventid() {
     return eventid;
   }
 
-  public void setEventid(Integer eventid) {
+  public void setEventid(Long eventid) {
     this.eventid = eventid;
   }
