+++ src/main/java/com/venturenix/cmc/payload/response/EventUserResponse.java
@@ -9,8 +9,8 @@ public class EventUserResponse {
   private String token;
   private String type = "Bearer";
   private Long id; 
-  private Integer userid;
-  private Integer eventid;
+  private Long userid;
+  private Long eventid;
   private String status;
   
   private LocalDateTime createddate;
@@ -25,7 +25,7 @@ public class EventUserResponse {
   private Integer updatedby;
   
   
-  public EventUserResponse(String accessToken, Long id, Integer userid, Integer eventid, String name, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
+  public EventUserResponse(String accessToken, Long id, Long userid, Long eventid, String name, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
     this.token = accessToken;
     this.id = id;
     this.userid = userid;
@@ -62,19 +62,19 @@ public class EventUserResponse {
     this.id = id;
   }
 
-  public Integer getUserid() {
+  public Long getUserid() {
     return userid;
   }
 
-  public void setUserid(Integer userid) {
+  public void setUserid(Long userid) {
     this.userid = userid;
   }  
 
-  public Integer getEventid() {
+  public Long getEventid() {
     return eventid;
   }
 
-  public void setEventid(Integer eventid) {
+  public void setEventid(Long eventid) {
     this.eventid = eventid;
   }
