+++ src/main/java/com/venturenix/cmc/payload/response/GroupQuestionSubmitResponse.java
@@ -10,16 +10,16 @@ public class GroupQuestionSubmitResponse {
   private String type = "Bearer";
   private Long id; 
 
-  private Integer eventid;
+  private Long eventid;
 
 
-  private Integer groupid;
+  private Long groupid;
 
 
-  private Integer questionid;  
+  private Long questionid;  
 
 
-  private Integer userid;  
+  private Long userid;  
 
 
   private Double run_time_by_sec;  
@@ -40,7 +40,7 @@ public class GroupQuestionSubmitResponse {
   private Integer updatedby;
   
   
-  public GroupQuestionSubmitResponse(String accessToken, Long id, Integer eventid, Integer groupid, Integer questionid, Integer userid, Double run_time_by_sec, LocalDateTime submit_time, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
+  public GroupQuestionSubmitResponse(String accessToken, Long id, Long eventid, Long groupid, Long questionid, Long userid, Double run_time_by_sec, LocalDateTime submit_time, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
     this.token = accessToken;
     this.id = id;
     this.eventid = eventid;
@@ -82,35 +82,35 @@ public class GroupQuestionSubmitResponse {
     this.id = id;
   }
 
-    public Integer getEventid() {
+    public Long getEventid() {
     return eventid;
   }
 
-  public void setEventid(Integer eventid) {
+  public void setEventid(Long eventid) {
     this.eventid = eventid;
   }
 
-  public Integer getGroupid() {
+  public Long getGroupid() {
     return groupid;
   }
 
-  public void setGroupid(Integer groupid) {
+  public void setGroupid(Long groupid) {
     this.groupid = groupid;
   }  
 
-  public Integer getQuestionid() {
+  public Long getQuestionid() {
     return questionid;
   }
 
-  public void setQuestionid(Integer questionid) {
+  public void setQuestionid(Long questionid) {
     this.questionid = questionid;
   }
 
-  public Integer getUserid() {
+  public Long getUserid() {
     return userid;
   }
 
-  public void setUserid(Integer userid) {
+  public void setUserid(Long userid) {
     this.userid = userid;
   }
