+++ src/main/java/com/venturenix/cmc/payload/response/UserScoreResponse.java
@@ -8,15 +8,15 @@ import java.util.Date;
 public class UserScoreResponse {
   private String token;
   private String type = "Bearer";
-  private Integer id; 
+  private Long id; 
  
-  private Integer eventid;
+  private Long eventid;
 
  
-  private Integer userid;
+  private Long userid;
 
  
-  private Integer questionid;
+  private Long questionid;
 
  
   private Integer testcase_pass_total;
@@ -41,7 +41,7 @@ public class UserScoreResponse {
   private Integer updatedby;
   
   
-  public UserScoreResponse(String accessToken, Integer id, Integer eventid, Integer userid, Integer questionid, Integer testcase_total, Integer testcase_pass_total, Double testcase_score_total, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
+  public UserScoreResponse(String accessToken, Long id, Long eventid, Long userid, Long questionid, Integer testcase_total, Integer testcase_pass_total, Double testcase_score_total, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
     this.token = accessToken;
     this.id = id;
     this.eventid = eventid;
@@ -74,35 +74,35 @@ public class UserScoreResponse {
     this.type = tokenType;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-   public Integer getEventid() {
+   public Long getEventid() {
     return eventid;
   }
 
-  public void setEventid(Integer eventid) {
+  public void setEventid(Long eventid) {
     this.eventid = eventid;
   }  
 
-  public Integer getUserid() {
+  public Long getUserid() {
     return userid;
   }
 
-  public void setUserid(Integer userid) {
+  public void setUserid(Long userid) {
     this.userid = userid;
   }
 
-  public Integer getQuestionid() {
+  public Long getQuestionid() {
     return questionid;
   }
 
-  public void setQuestionid(Integer questionid) {
+  public void setQuestionid(Long questionid) {
     this.questionid = questionid;
   }
