+++ src/main/java/com/venturenix/cmc/repository/UserQuestionSubmitRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface UserQuestionSubmitRepository extends JpaRepository<UserQuestionSubmit, Long> {
 
-    Optional<UserQuestionSubmit> findById(int id);
+    Optional<UserQuestionSubmit> findById(Long id);
 
         List<UserQuestionSubmit> findAll();
